delete goal

+++ backend/controllers/goalController.js
@@ -71,6 +71,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  const user = await User.findById(req.user.id);
+
+  // Check for User
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches in goal
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   await Goal.findByIdAndDelete(req.params.id);
 
   res.status(200).json({ message: "Goal deleted successfully" });
